refactor(e2e): move extends-up to its own directory
Separate labels into separate directories for clarity:
- e2e/tests/extends/ — config-level read-configuration tests
- e2e/tests/extends-up/ — container up test with Docker provider
