+++ e2e/tests/extends-up/testdata/up-extends/.devcontainer/base.json
