+++ e2e/tests/extends-up/testdata/up-extends/.devcontainer/devcontainer.json
